requirements.txt
discord.py
requests
pytz
pymongo[srv] >= 4.11.3
boto3
dotenv